ci: updates versions, starts caching aztec image, node modules, artifacts and more (#16)
This PR aims to:
- Update all jobs versions to their latest
- Cache everywhere it makes sense (yarn global, node_modules, noir
compilations and artifacts)
- We remove the double-installation of aztec images which took a lot of
time: cli and set-up aztec version. We can directly install cli with the
aztec version and save ~30s on tests.

I think the only highly opinionated thing is that I've made the
compare-benchmark skip the job if no noir files, nargo config or aztec
version changed. LMK if you disagree and I'll roll it back.

---------

Co-authored-by: xorsal <zorzal@wonderland.xyz>